[225701, 245847, 246188] documentation, remove warnings, move non API packages to under internal

+++ plugins/org.eclipse.actf.visualization.engines.lowvision/src/org/eclipse/actf/visualization/internal/engines/lowvision/DecisionMaker.java
@@ -545,7 +545,7 @@ public class DecisionMaker {
 		}
 
 		// check problem type
-		short type1 = _p1.getLowVisionProblemType();
+		int type1 = _p1.getLowVisionProblemType();
 		if (type1 != _p2.getLowVisionProblemType()) {
 			return (false);
 		}
