+++ plugins/org.eclipse.actf.visualization.engines.lowvision/src/org/eclipse/actf/visualization/internal/engines/lowvision/LowVisionProblemConverter.java
@@ -60,7 +60,7 @@ public class LowVisionProblemConverter {
     }
 
     private static String getLVProblemColorString(LowVisionProblemGroup problem, boolean isFore) {
-        short probType;
+        int probType;
         int origAll;
         int origR;
         int origG;
